장바구니 수정

+++ sintoburi/src/main/java/com/kh/sintoburi/controller/gr/BasketController.java
@@ -1,6 +1,7 @@
 package com.kh.sintoburi.controller.gr;
 
 import java.util.List;
+import java.util.Map;
 
 import javax.servlet.http.HttpSession;
 
@@ -9,8 +10,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.kh.sintoburi.domain.gr.BasketDetailDto;
 import com.kh.sintoburi.service.gr.BasketService;
@@ -21,6 +24,7 @@ public class BasketController {
 	@Autowired
 	private BasketService basketService;
 	
+	// 장바구니 목록 보기
 	@GetMapping("/list")
 	public void main(Model model, HttpSession session) {
 		
@@ -30,23 +34,36 @@ public class BasketController {
 		model.addAttribute("list", list);
 	}
 		
-	@PostMapping("/Modify")
-		public boolean basketModify(int p_count, int bdno, Model model) {
-					
-		boolean result = basketService.modifyCount(p_count, bdno);
-		//if
-		
+	// 개수 수정
+	@PostMapping("/modCount")
+	@ResponseBody
+	public boolean modCount(@RequestBody Map<String, Integer> map) {
+		System.out.println("modCount...map:" + map);
+		int bdno = map.get("bdno");
+		int p_count = map.get("p_count");
+		System.out.println("modCount...bdno:" + bdno);
+		System.out.println("modCount...p_count:" + p_count);
+		boolean result = false;
+		if(p_count == 0)	{
+			result = basketService.removeOne(bdno);
+		} else {
+			result = basketService.modifyCount(p_count, bdno);
+		}
 		return result;
 	}
 	
-	@PostMapping("/RemoveOne")
-	public void basketRemoveOne(@RequestParam("bdno") int bdno) {
-		basketService.removeOne(bdno);
+	@PostMapping("/removeOne")
+	@ResponseBody
+	public void removeOne(@RequestParam("bdno") int bdno) {
+		System.out.println("removeOne...");
+		boolean result = basketService.removeOne(bdno);
+		
 	}
 	
-	@PostMapping("/RemoveAll")
-	public void basketRemoveAll(@RequestParam("pno") int pno) {
-		basketService.removeAll(pno);
+	@PostMapping("/removeAll")
+	public void removeAll(@RequestParam("pno") int pno) {
+		boolean result = basketService.removeAll(pno);
+		
 	}
